feat:  use es module, drop Node <  18
BREAKING CHANGE: use es module, drop Node <  18
