+++ README.md
@@ -5,8 +5,9 @@ Get stdout window width, with four fallbacks, `tty`, `output.columns`, a custom 
 [![npm version](https://badge.fury.io/js/cli-width.svg)](http://badge.fury.io/js/cli-width)
 [![Coverage Status](https://coveralls.io/repos/knownasilya/cli-width/badge.svg?branch=master&service=github)](https://coveralls.io/github/knownasilya/cli-width?branch=master)
 
-Tested against Node v12 to v20.
+Tested against Node v18 to v21.
 Includes TypeScript types.
+Uses ES Modules.
 
 ## Usage
 
@@ -15,9 +16,10 @@ npm install --save cli-width
 ```
 
 ```js
-const cliWidth = require('cli-width');
+import cliWidth from 'cli-width';
 
-cliWidth(); // maybe 204 :)
+const width = cliWidth();
+console.log(width); // maybe 204 :)
 ```
 
 You can also set the `CLI_WIDTH` environment variable.
@@ -27,7 +29,7 @@ the default width value is going to be `0`, that can be changed using the config
 
 ## API
 
-### cliWidth([options])
+### cliWidth([options]): number
 
 `cliWidth` can be configured using an `options` parameter, the possible properties are:
 
@@ -35,13 +37,15 @@ the default width value is going to be `0`, that can be changed using the config
 - **output**\<object\> A stream to be used to read width values from, defaults to `process.stdout`
 - **tty**\<object\> TTY module to try to read width from as a fallback, defaults to `require('tty')`
 
+Returns a `number`
+
 ### Examples
 
 Defining both a default width value and a stream output to try to read from:
 
 ```js
-const cliWidth = require('cli-width');
-const ttys = require('ttys');
+import cliWidth from 'cli-width';
+import ttys from 'ttys';
 
 cliWidth({
   defaultWidth: 80,
@@ -52,8 +56,8 @@ cliWidth({
 Defines a different tty module to read width from:
 
 ```js
-const cliWidth = require('cli-width');
-const ttys = require('ttys');
+import cliWidth from 'cli-width';
+import ttys from 'ttys';
 
 cliWidth({
   tty: ttys,
