+++ package.json
@@ -2,10 +2,10 @@
   "name": "cli-width",
   "version": "4.1.0",
   "description": "Get stdout window width, with two fallbacks, tty and then a default.",
-  "main": "index.js",
+  "main": "index.mjs",
   "scripts": {
-    "test": "node test | tspec",
-    "coverage": "nyc node test | tspec",
+    "test": "node test/index.mjs | tspec",
+    "coverage": "nyc node test/index.mjs | tspec",
     "coveralls": "npm run coverage -s && coveralls < coverage/lcov.info",
     "release": "standard-version"
   },
@@ -20,18 +20,17 @@
   },
   "homepage": "https://github.com/knownasilya/cli-width",
   "engines": {
-    "node": ">= 12"
+    "node": ">= 18"
   },
   "devDependencies": {
     "coveralls": "^3.1.1",
     "nyc": "^15.1.0",
     "standard-version": "^9.3.2",
     "tap-spec": "^5.0.0",
-    "tape": "^5.5.2"
+    "tape": "^5.7.4"
   },
   "volta": {
-    "node": "12.22.11",
-    "npm": "8.5.5"
+    "node": "18.19.0"
   },
   "files": [
     "index.js",
