+++ test/index.mjs
@@ -1,9 +1,6 @@
-'use strict';
-
-var tty = require('tty');
-var test = require('tape');
-var lib = require('../');
-
+import tty from 'node:tty';
+import test from 'tape';
+import lib from '../index.mjs';
 
 test('uses process.stdout.getWindowSize', function (t) {
   // mock stdout.getWindowSize
@@ -22,7 +19,7 @@ test('uses defaultWidth if process.stdout.getWindowSize reports width of 0', fun
 
   t.equal(lib({ defaultWidth: 10 }), 10, 'equal to mocked, 10');
   t.end();
-})
+});
 
 test('uses tty.getWindowSize', function (t) {
   process.stdout.getWindowSize = undefined;
@@ -42,10 +39,11 @@ test('uses default if tty.getWindowSize reports width of 0', function (t) {
 
   t.equal(lib({ defaultWidth: 10 }), 10, 'equal to mocked, 10');
   t.end();
-})
+});
 
 test('uses custom env var', function (t) {
-  var oldWidth = process.stdout.columns;
+  let oldWidth = process.stdout.columns;
+
   process.stdout.columns = undefined;
   tty.getWindowSize = undefined;
   process.env.CLI_WIDTH = 30;
@@ -58,7 +56,8 @@ test('uses custom env var', function (t) {
 });
 
 test('uses default if env var is not a number', function (t) {
-  var oldWidth = process.stdout.columns;
+  let oldWidth = process.stdout.columns;
+
   process.stdout.columns = undefined;
   process.env.CLI_WIDTH = 'foo';
 
@@ -70,7 +69,8 @@ test('uses default if env var is not a number', function (t) {
 });
 
 test('uses default', function (t) {
-  var oldWidth = process.stdout.columns;
+  let oldWidth = process.stdout.columns;
+
   process.stdout.columns = undefined;
   tty.getWindowSize = undefined;
 
@@ -81,7 +81,8 @@ test('uses default', function (t) {
 });
 
 test('uses overridden default', function (t) {
-  var oldWidth = process.stdout.columns;
+  let oldWidth = process.stdout.columns;
+
   process.stdout.columns = undefined;
 
   t.equal(lib({ defaultWidth: 10 }), 10, 'user-set defaultWidth value, 10');
@@ -91,10 +92,10 @@ test('uses overridden default', function (t) {
 });
 
 test('uses user-configured output stream', function (t) {
-  var outputMock = {
-      getWindowSize: function () {
-          return [10];
-      }
+  const outputMock = {
+    getWindowSize: function () {
+      return [10];
+    },
   };
 
   t.equal(lib({ output: outputMock }), 10, 'user-set output stream, 10');
@@ -103,10 +104,10 @@ test('uses user-configured output stream', function (t) {
 });
 
 test('uses user-configured tty', function (t) {
-  var ttyMock = {
-      getWindowSize: function () {
-          return [2, 5];
-      }
+  const ttyMock = {
+    getWindowSize: function () {
+      return [2, 5];
+    },
   };
 
   t.equal(lib({ tty: ttyMock }), 5, 'user-set tty, 5');
@@ -115,7 +116,8 @@ test('uses user-configured tty', function (t) {
 });
 
 test('uses output.columns', function (t) {
-  var oldWidth = process.stdout.columns;
+  let oldWidth = process.stdout.columns;
+
   process.stdout.columns = 15;
   process.stdout.getWindowSize = undefined;
   delete process.env.CLI_WIDTH;
@@ -124,4 +126,4 @@ test('uses output.columns', function (t) {
 
   process.stdout.columns = oldWidth;
   t.end();
-})
+});
